Splitting structure and parameter estimators into their own directories
